// libs/engine/src/dynamic-libs/vk/enums/Vk_StructureType.ts
/**
 * Structure type enumerant.
 *
 * @see https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/VkStructureType.html
 */
export enum Vk_StructureType {
  // MARK: VK 1.0
  APPLICATION_INFO = 0,
  INSTANCE_CREATE_INFO = 1,
  DEVICE_QUEUE_CREATE_INFO = 2,
  DEVICE_CREATE_INFO = 3,
  SUBMIT_INFO = 4,
  MEMORY_ALLOCATE_INFO = 5,
  MAPPED_MEMORY_RANGE = 6,
  BIND_SPARSE_INFO = 7,
  FENCE_CREATE_INFO = 8,
  SEMAPHORE_CREATE_INFO = 9,
  EVENT_CREATE_INFO = 10,
  QUERY_POOL_CREATE_INFO = 11,
  BUFFER_CREATE_INFO = 12,
  BUFFER_VIEW_CREATE_INFO = 13,
  IMAGE_CREATE_INFO = 14,
  IMAGE_VIEW_CREATE_INFO = 15,
  SHADER_MODULE_CREATE_INFO = 16,
  PIPELINE_CACHE_CREATE_INFO = 17,
  PIPELINE_SHADER_STAGE_CREATE_INFO = 18,
  PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO = 19,
  PIPELINE_INPUT_ASSEMBLY_STATE_CREATE_INFO = 20,
  PIPELINE_TESSELLATION_STATE_CREATE_INFO = 21,
  PIPELINE_VIEWPORT_STATE_CREATE_INFO = 22,
  PIPELINE_RASTERIZATION_STATE_CREATE_INFO = 23,
  PIPELINE_MULTISAMPLE_STATE_CREATE_INFO = 24,
  PIPELINE_DEPTH_STENCIL_STATE_CREATE_INFO = 25,
  PIPELINE_COLOR_BLEND_STATE_CREATE_INFO = 26,
  PIPELINE_DYNAMIC_STATE_CREATE_INFO = 27,
  GRAPHICS_PIPELINE_CREATE_INFO = 28,
  COMPUTE_PIPELINE_CREATE_INFO = 29,
  PIPELINE_LAYOUT_CREATE_INFO = 30,
  SAMPLER_CREATE_INFO = 31,
  DESCRIPTOR_SET_LAYOUT_CREATE_INFO = 32,
  DESCRIPTOR_POOL_CREATE_INFO = 33,
  DESCRIPTOR_SET_ALLOCATE_INFO = 34,
  WRITE_DESCRIPTOR_SET = 35,
  COPY_DESCRIPTOR_SET = 36,
  FRAMEBUFFER_CREATE_INFO = 37,
  RENDER_PASS_CREATE_INFO = 38,
  COMMAND_POOL_CREATE_INFO = 39,
  COMMAND_BUFFER_ALLOCATE_INFO = 40,
  COMMAND_BUFFER_INHERITANCE_INFO = 41,
  COMMAND_BUFFER_BEGIN_INFO = 42,
  RENDER_PASS_BEGIN_INFO = 43,
  BUFFER_MEMORY_BARRIER = 44,
  IMAGE_MEMORY_BARRIER = 45,
  MEMORY_BARRIER = 46,

  // MARK: VK 1.0 - KHR Extensions
  SWAPCHAIN_CREATE_INFO_KHR = 1000001000,
  PRESENT_INFO_KHR = 1000001001,
  XLIB_SURFACE_CREATE_INFO_KHR = 1000004000,
  XCB_SURFACE_CREATE_INFO_KHR = 1000005000,
  WAYLAND_SURFACE_CREATE_INFO_KHR = 1000006000,
  WIN32_SURFACE_CREATE_INFO_KHR = 1000009000,

  // MARK: VK 1.1
  QUEUE_FAMILY_PROPERTIES_2 = 1000059005,

  // MARK: VK 1.1 - EXT Extensions
  METAL_SURFACE_CREATE_INFO_EXT = 1000217000,
}